Improve `tasks` namespace
Make file naming more logical.
