+++ src/tasks/__init__.py
@@ -1,4 +1,6 @@
 from . import check_software_dependencies
+from . import download_version
 from . import find_versions
-from . import installation
-from . import manage_local_files
+from . import install_version
+from . import migrate_config
+from . import remove_version
