+++ src/tasks/download_version.py
@@ -1,5 +1,4 @@
 import os
-import shutil
 from src import Version, utils
 
 
@@ -24,19 +23,3 @@ def download_version(version: Version) -> None:
         cwd=os.path.join(pyra_dir, "ui-installers"),
     )
     utils.shell_utils.pretty_print("Downloaded UI", color="green")
-
-
-def remove_version(version: Version) -> None:
-    """For a given release version "x.y.z", remove the code and its ui-installer."""
-
-    pyra_dir = os.path.join(utils.directory_utils.get_documents_dir(), "pyra")
-    ui_installer_path = os.path.join(
-        pyra_dir, "ui-installers", f"Pyra.UI_{version.as_str()}_x64_en-US.msi"
-    )
-    code_dir = os.path.join(pyra_dir, f"pyra-{version.as_str()}")
-
-    if os.path.isdir(code_dir):
-        shutil.rmtree(code_dir)
-
-    if os.path.isfile(ui_installer_path):
-        os.remove(ui_installer_path)
