+++ src/utils/__init__.py
@@ -2,9 +2,8 @@ from .directory_utils import (
     get_desktop_dir,
     get_documents_dir,
     initialize_pyra_directories,
+    pyra_dir_is_in_env_path,
 )
-
-from .migration_utils import migrate_config
 
 from .shell_utils import (
     run_shell_command,
